weather-app/weather-app-backend/requirements.txt
Flask
gunicorn
flask-cors
CORS
requests
load_dotenv